separated LoginModal from LoginButton
